begin porting filesystem functions from xygine

+++ crogine/src/ecs/systems/ModelRenderer.cpp
@@ -90,13 +90,13 @@ void ModelRenderer::process(Time)
             {
                 if (model.m_materials[i].blendMode != Material::BlendMode::None)
                 {
-                    transparent.second.matIDs.push_back(i);
+                    transparent.second.matIDs.push_back(static_cast<std::int32_t>(i));
                     transparent.second.flags = static_cast<int64>(worldPos.z * 1000000.f); //suitably large number to shift decimal point
                     transparent.second.flags += 0x0FFF000000000000; //gaurentees embiggenment so that sorting places transparent last
                 }
                 else
                 {
-                    opaque.second.matIDs.push_back(i);
+                    opaque.second.matIDs.push_back(static_cast<std::int32_t>(i));
                     opaque.second.flags = static_cast<int64>(-worldPos.z * 1000000.f);
                 }
             }
